plugins/plugin-gdb-openocd/che-plugin-gdb-openocd-server/src/test/java/org/eclipse/che/plugin/gdbopenocd/server/parser/GdbOpenocdPTypeTest.java
/*******************************************************************************
 * Copyright (c) 2012-2016 Codenvy, S.A.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v1.0
 * which accompanies this distribution, and is available at
 * http://www.eclipse.org/legal/epl-v10.html
 *
 * Contributors:
 *   Codenvy, S.A. - initial API and implementation
 *******************************************************************************/
package org.eclipse.che.plugin.gdbopenocd.server.parser;

import org.testng.annotations.Test;

import static org.testng.Assert.assertEquals;

/**
 * @author Anatoliy Bazko
 */
public class GdbOpenocdPTypeTest {

    @Test
    public void testParse() throws Exception {
        GdbOpenocdOutput gdbOpenocdOutput = GdbOpenocdOutput.of("type = int\n");

        GdbOpenocdPType gdbOpenocdPType = GdbOpenocdPType.parse(gdbOpenocdOutput);

        assertEquals(gdbOpenocdPType.getType(), "int");
    }
}
